VM error line tracking is now usize
The number of lines can be as many as the operating system can count.
This allows files with lots of empty lines and comments, which don't
count towards the capped instruction memory of the hack computer.

+++ hack_virtual_machine/src/lib.rs
@@ -49,13 +49,13 @@ impl std::str::FromStr for Segment {
 #[derive(Debug)]
 pub enum Error {
     /// The command has too few, too many, or wrong type of arguments
-    InvalidArgs(u16),
+    InvalidArgs(usize),
     /// Upstream IO error.
     Io(std::io::Error),
     /// The command is not part of the VM spec
-    UnknownCommand(u16),
+    UnknownCommand(usize),
     /// The segement is not part of the VM spec
-    UnknownSegment(u16),
+    UnknownSegment(usize),
 }
 
 impl std::fmt::Display for Error {
